Check if an object is a numeric vector of valid logarithmic Bayes factor values

inspect_log_bf checks if an object is a numeric vector of valid logarithmic Bayes factor values. This can be useful to validate inputs, intermediate calculations or outputs in user-defined functions.

inspect_log_bf(x, allow_nas = TRUE, warning_nas = TRUE)

Arguments

x
An arbitrary object.

allow_nas
Logical value. If TRUE then NA and NaN values in x are allowed. If FALSE, execution is stopped and an error message is thrown in case there are NA or NaN values in x.

warning_nas
Logical value. If TRUE then the presence of NA or NaN values in x generates a warning message. NA and NaN pass silently otherwise (if allow_nas is TRUE).

Value

inspect_log_bf does not return any output. There are three possible outcomes:

The call is silent if:

x is a numeric vector of valid logarithmic Bayes factor values without NA or NaN values.

x is a numeric vector of valid logarithmic Bayes factor values with some NA or NaN values, allow_nas is set to TRUE and warning_nas is set to FALSE.

An informative warning message is given if x is a numeric vector of valid logarithmic Bayes factor values, there are NA or NaN values in x and both allow_nas and warning_nas are set to TRUE.

An informative error message is thrown and the execution is stopped if:

x is not a numeric vector of valid logarithmic Bayes factor values.

x is a numeric vector of valid logarithmic Bayes factor values with some NA or NaN values and allow_nas is set to FALSE.

Details

inspect_log_bf conducts a series of tests to check if x is a numeric vector of valid logarithmic Bayes factor values. Namely, inspect_log_bf checks if:

x is NULL or empty.

x is a numeric, atomic vector.

x has NA or NaN values.

See also

Examples

# Calls that pass silently:
x1 <- c(0, 0.5, 1, 10, 50, 100)
x2 <- c(NA, 0.5, 1, 10, 50, 100)
inspect_log_bf(x1)
inspect_log_bf(x2, warning_nas = FALSE)
inspect_log_bf(x2, allow_nas = TRUE, warning_nas = FALSE)

# Call that throws an informative warning message:
if (FALSE) y <- c(0.1, 0.2, NA, 0.4, 0.5)
if (FALSE) inspect_log_bf(y)
if (FALSE) inspect_log_bf(y, warning_nas = TRUE)
if (FALSE) inspect_log_bf(y, allow_nas = TRUE, warning_nas = TRUE)

# Calls that throw informative error messages:
if (FALSE) mylist <- list(NULL, TRUE, factor(.5), matrix(0.5),
         "0.5", list(0.5), numeric(0), NA, NaN)
if (FALSE) inspect_log_bf(mylist[[1]])
if (FALSE) inspect_log_bf(mylist[[2]])
if (FALSE) inspect_log_bf(mylist[[3]])
if (FALSE) inspect_log_bf(mylist[[4]])
if (FALSE) inspect_log_bf(mylist[[5]])
if (FALSE) inspect_log_bf(mylist[[6]])
if (FALSE) inspect_log_bf(mylist[[7]])
if (FALSE) inspect_log_bf(mylist[[8]])
if (FALSE) inspect_log_bf(mylist[[9]])